# Query improvements for RosterEntry.students_not_on_team