## What

- Use `flat_map` instead of `map(...).flatten`, reduces number of copies of arrays
- Use `user_id` instead of `user.id`, removes the n+1 query
- Remove raw SQL and replace with ActiveRecord syntax for readability

Some Benchmark stats:

Old Code:

~~~ruby
[68] pry(main)> Benchmark.ms { RosterEntry.students_not_on_team_old(group_assignment.reload).to_a; }
  GroupAssignment Load (4.2ms)  SELECT  "group_assignments".* FROM "group_assignments" WHERE "group_assignments"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
  GroupAssignmentRepo Load (1.8ms)  SELECT "group_assignment_repos".* FROM "group_assignment_repos" WHERE "group_assignment_repos"."group_assignment_id" = $1  [["group_assignment_id", 1]]
  RepoAccess Load (8.3ms)  SELECT "repo_accesses".* FROM "repo_accesses" INNER JOIN "groups_repo_accesses" ON "repo_accesses"."id" = "groups_repo_accesses"."repo_access_id" INNER JOIN "groups" ON "groups_repo_accesses"."group_id" = "groups"."id" WHERE "groups"."id" = $1  [["id", 1]]
  User Load (2.0ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = $1 LIMIT $2  [["id", 2], ["LIMIT", 1]]
  User Load (1.7ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = $1 LIMIT $2  [["id", 3], ["LIMIT", 1]]
  RosterEntry Load (4.9ms)  SELECT "roster_entries".* FROM "roster_entries" WHERE (roster_entries.user_id IS NULL OR
roster_entries.user_id NOT IN (2,3)
)
=> 29.058000014629215
~~~


New Code:

~~~ruby
[69] pry(main)> Benchmark.ms { RosterEntry.students_not_on_team(group_assignment.reload).to_a; }
  GroupAssignment Load (1.7ms)  SELECT  "group_assignments".* FROM "group_assignments" WHERE "group_assignments"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
  GroupAssignmentRepo Load (1.6ms)  SELECT "group_assignment_repos".* FROM "group_assignment_repos" WHERE "group_assignment_repos"."group_assignment_id" = $1  [["group_assignment_id", 1]]
  RepoAccess Load (2.2ms)  SELECT "repo_accesses".* FROM "repo_accesses" INNER JOIN "groups_repo_accesses" ON "repo_accesses"."id" = "groups_repo_accesses"."repo_access_id" INNER JOIN "groups" ON "groups_repo_accesses"."group_id" = "groups"."id" WHERE "groups"."id" = $1  [["id", 1]]
  RosterEntry Load (4.1ms)  SELECT "roster_entries".* FROM "roster_entries" WHERE ("roster_entries"."user_id" IS NULL OR ("roster_entries"."user_id" NOT IN (2, 3)))
=> 13.928000000305474
~~~

We can notice that the new code has no n+1 queries, which results in improvement in query speed. Need to confirm this against a large `group_assignment` but expecting a __significant__ improvement.

This is another part of #1969 which discusses assignments, this pull request will hopefully improve  loading time of group assignments show page.