@shaunakpp 2019-08-08
This can be further improved by replacing the `NOT IN` clause with a 3-way `JOIN`, between `RepoAccess`, `GroupAssignmentRepo` and `RosterEntry`, but I need to further investigate on that. 😅